+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Unreleased
 
 * Adds BugReporting.setVideoRecordingFloatingButtonPosition API
-* Adds some missing string keys
+* Adds some missing string keys and their mappings on iOS and Android
+* Fixes an issue with some string keys not working on Android
 
 ## 11.2.0 (2022-09-08)